+++ lib/util/direction_animations/simple_direction_animation.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 import 'dart:ui';
 
-import 'package:bonfire/base/bonfire_game_interface.dart';
 import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/util/sprite_animation_render.dart';