+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# NEXT
+- Fix bug unwanted push when to component have BlockMovementCollision.
+
 ## 3.0.7
 - adds set `mapZoomFit` in `BonfireCamera`.
 - adds set `moveOnlyMapArea` in `BonfireCamera`.